Fix bpf deadlock, which might occur when using DHCP
It seems like it's possible that during the processing of bpf_callback
sometimes set_bpf_tap is called which then deadlocks BPF
in the kernel.

This manifests in different problems on the network stack and OSX
can't even shutdown anymore because it can't unload the KEXT.

set_bpf_tap is deprecated so we could get rid of it, but I don't know
since what kernel version the new implementation is available.